Set up authentication for editing user account, changed the functions for editing user. Updated navbar

+++ Project01/default.aspx.cs
@@ -12,6 +12,11 @@ using Project01.Models;
 using System.Web.ModelBinding;
 using System.Linq.Dynamic;
 
+//required for Identity and OWIN Security
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.EntityFramework;
+using Microsoft.Owin.Security;
+
 /**
  * @author: Matthew Lane
  * @date: June 14, 2016
@@ -39,7 +44,7 @@ namespace Project01
             if (!IsPostBack)
             {
                 // Get the game data
-                this.GetGames(1);
+                this.GetGames();
             }
         }
 
@@ -53,7 +58,7 @@ namespace Project01
     */
         protected void GetDates()
         {
-            using (FootballScoreModel db = new FootballScoreModel())
+            using (Football_Model db = new Football_Model())
             {
                 // query the Students Table using EF and LINQ
                 var Games = (from allGames in db.Football_Scores where allGames.GameID == 1 select allGames);
@@ -69,22 +74,23 @@ namespace Project01
          * @returns {void}
          * @param int Week
          */
-        protected void GetGames(int Week)
+        protected void GetGames()
         {
-           // if (Week == 1)
 
             // connect to EF
-            using (FootballScoreModel db = new FootballScoreModel())
+            using (Football_Model db = new Football_Model())
             {
 
+                int gameWeek = Convert.ToInt32(GameWeekDropDownList.SelectedValue);
+
                    // query the Students Table using EF and LINQ
-                   var GameNum = (from allGames in db.Football_Scores
+                var GameNum = (from allGames in db.Football_Scores
                                     select allGames.GameID);
 
                     // query the Students Table using EF and LINQ
                     var Games = (from allGames in db.Football_Scores
-                            // where allGames.GameID == 1
-                             select allGames);
+                                 where allGames.GameWeek == gameWeek
+                                 select allGames);
 
                 // bind the result to the GridView
                 GamesGridView.DataSource = Games.ToList();
@@ -107,28 +113,38 @@ namespace Project01
          
         protected void GamesGridView_RowDeleting(object sender, GridViewDeleteEventArgs e)
         {
-            // store which row was clicked
-            int selectedRow = e.RowIndex;
 
-            // get the selected StudentID using the Grid's DataKey collection
-            int GameID = Convert.ToInt32(GamesGridView.DataKeys[selectedRow].Values["GameID"]);
+            var authenticationManager = HttpContext.Current.GetOwinContext().Authentication;
 
-            // use EF to find the selected student in the DB and remove it
-            using (FootballScoreModel db = new FootballScoreModel())
+            if (authenticationManager.User.Identity.Name == "")
             {
-                // create object of the Student class and store the query string inside of it
-                Football_Score deletedGame = (from savedGames in db.Football_Scores
-                                              where savedGames.GameID == GameID
-                                          select savedGames).FirstOrDefault();
+                Response.Redirect("~/Login.aspx");
+            }
+            else
+            {
+                // store which row was clicked
+                int selectedRow = e.RowIndex;
 
-                // remove the selected student from the db
-                db.Football_Scores.Remove(deletedGame);
+                // get the selected StudentID using the Grid's DataKey collection
+                int GameID = Convert.ToInt32(GamesGridView.DataKeys[selectedRow].Values["GameID"]);
 
-                // save my changes back to the database
-                db.SaveChanges();
+                // use EF to find the selected student in the DB and remove it
+                using (Football_Model db = new Football_Model())
+                {
+                    // create object of the Student class and store the query string inside of it
+                    Football_Score deletedGame = (from savedGames in db.Football_Scores
+                                                  where savedGames.GameID == GameID
+                                                  select savedGames).FirstOrDefault();
 
-                // refresh the grid
-                this.GetGames(1);
+                    // remove the selected student from the db
+                    db.Football_Scores.Remove(deletedGame);
+
+                    // save my changes back to the database
+                    db.SaveChanges();
+
+                    // refresh the grid
+                    this.GetGames();
+                }
             }
         }
 
@@ -148,26 +164,7 @@ namespace Project01
             GamesGridView.PageIndex = e.NewPageIndex;
 
             // refresh the grid
-            this.GetGames(1);
-        }
-
-        /**
-        * <summary>
-        * This event handler allows for the number of games being displayed on one page to be changed
-        * </summary>
-        * 
-        * @method PageSizeDropDownList_SelectedIndexChanged
-        * @param {object} sender
-        * @param {GridViewPageEventArgs} e
-        * @returns {void}
-        */
-        protected void PageSizeDropDownList_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            // Set the new Page size
-            GamesGridView.PageSize = Convert.ToInt32(PageSizeDropDownList.SelectedValue);
-
-            // refresh the grid
-            this.GetGames(Convert.ToInt32(PageSizeDropDownList.SelectedValue));
+            this.GetGames();
         }
 
         /**
@@ -186,10 +183,26 @@ namespace Project01
             Session["SortColumn"] = e.SortExpression;
 
             // Refresh the Grid
-            this.GetGames(1);
+            this.GetGames();
 
             // toggle the direction
             Session["SortDirection"] = Session["SortDirection"].ToString() == "ASC" ? "DESC" : "ASC";
+        }
+
+
+        /**
+        * <summary>
+        * This event handler allows speficying which week to display games from
+        * </summary>
+        * 
+        * @method GameWeekDropDownList_SelectedIndexChanged
+        * @param {object} sender
+        * @param {GridViewPageEventArgs} e
+        * @returns {void}
+        */
+        protected void GameWeekDropDownList_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            GetGames();
         }
     }
 }
